Make Hash field calculations internal responsibility
The Hash is now calculated internally by the client instead of requiring
the caller to use the helper function to calculate the Hash before
passing the message to the client.
