+++ client/types/startsettings.go
@@ -30,6 +30,8 @@ type StartSettings struct {
 	// The remote config status. If nil is passed it will force
 	// the Server to send a remote config back. It is not required to set the Hash
 	// field, it will be calculated by Start() function.
+	// The Hash field will be calculated and updated from the content of the rest of
+	// the fields.
 	RemoteConfigStatus *protobufs.RemoteConfigStatus
 
 	LastConnectionSettingsHash []byte
